ai/requirements.txt
# Core Frameworks
fastapi
uvicorn

# Data Validation
pydantic

# Google Gemini API
google-genai

# Environment Management
python-dotenv

# Solana / Anchor
anchorpy
solana
solders
httpx
construct